requirements.txt
qdrant-client
tqdm
google-genai
markdownify
requests
gradio
python-dotenv